requirements.txt
pathway[xpack]>=0.14.0
numpy>=1.26.0
fastapi==0.110.0
uvicorn==0.29.0
google-generativeai==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
pytz>=2024.1
